.amazon-project/requirements_production.txt
streamlit>=1.28.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
pathlib2>=2.3.0
pandas>=2.0.0
tensorflow>=2.13.0
